requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.8"
Django==4.0.6
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
requests==2.27.1
